+++ nimjl/conversions/dict_tuples.nim
@@ -1,20 +1,22 @@
-import ../cores
 import ../types
 import ../functions
+import ../cores
 
 import std/tables
 import std/strutils
 import std/strformat
 
 proc jlTupleToNim*(val: JlValue, tup: var tuple) =
-  # collect(keys(val))
+  # julia> collect(keys(val))
   var keys = jlCall("keys", val)
   keys = jlCall("collect", keys)
+
   # Tuple of JlSymbol
-  # length(collect(keys(val)))
+  # julia> length(collect(keys(val)))
   var show = getJlFunc("show")
   var sprint = getJlFunc("sprint")
   var nkeys = jlCall("length", keys).to(int)
+
   var i = 0
   for name, field in tup.fieldPairs:
     inc(i)
@@ -33,31 +35,28 @@ proc jlTupleToNim*(val: JlValue, tup: var tuple) =
       raise newException(JlError, "Tuple conversion from Julia to Nim failed ! Fields must identical")
 
 proc jlDictToNim*[U, V: string|SomeNumber|bool](val: JlValue, tab: var Table[U, V]) =
-  # collect(keys(val))
+  # julia> collect(keys(val))
   var keys = jlCall("keys", val)
   keys = jlCall("collect", keys)
+
   # Tuple of JlSymbol
-  # length(collect(keys(val)))
+  # julia> length(collect(keys(val)))
   var nkeys = jlCall("length", keys).to(int)
   for i in 1..nkeys:
     var key = jlCall("getindex", keys, i)
     var val = jlCall("getindex", val, key)
     tab[key.to(U)] = val.to(V)
 
-# Recursive import strategy
-import ./converttypes
-# Tuple helpers -> result is memory managed by Julia's GC
-# Convert object as tuple ?
 proc nimToUnnamedJlTuple(v: tuple): JlValue =
   var tupStr = $v
   result = jlEval(tupStr)
 
-proc nimToNamedJlTuple(v: tuple|object): JlValue =
+proc nimToNamedJlTuple*(v: tuple|object): JlValue =
   result = jlEval("NamedTuple()")
   for name, field in v.fieldPairs:
-    result = jlCall(JlBase, "setindex", result, toJlVal(field), jlSym(name))
+    result = jlCall(JlBase, "setindex", result, field, jlSym(name))
 
-proc isNamedTuple(v: tuple) : bool =
+proc isNamedTuple(v: tuple): bool =
   var i = 0
   for name, field in v.fieldPairs:
     if name != &"Field{i}":
@@ -71,11 +70,7 @@ proc nimToJlTuple*(v: tuple): JlValue =
   else:
     nimToUnnamedJlTuple(v)
 
-proc nimToJlTuple*(v: object): JlValue =
-  nimToNamedJlTuple(v)
-
 proc nimTableToJlDict*[U, V: string|SomeNumber](tab: Table[U, V]): JlValue =
   result = jlEval("Dict()")
   for name, field in tab:
-    discard jlCall(JlBase, "setindex!", result, toJlVal(field), name)
-
+    discard jlCall(JlBase, "setindex!", result, field, name)