+++ nimjl/conversions/fromjl.nim
@@ -1,13 +1,11 @@
 import ../private/jlcores
-import ../cores
 import ../types
 import ../arrays
 
-import ./boxunbox
+import ./unbox
 
 import std/json
 import std/tables
-import std/options
 import arraymancer
 
 {.push inline.}
@@ -23,6 +21,7 @@ proc toNimVal(x: JlValue, res: var string) =
 # Julia Tuple / Dict can't really be mapped to Nim's type so returning JsonNode is easier.
 # It introduces a "distinction" between to[T] -> T and to[T] -> JsonNode as return types
 # Forward declare for cyclic import
+proc toNimVal(x: JlValue, t: var object)
 proc toNimVal(x: JlValue, t: var tuple)
 proc toNimVal[U, V](x: JlValue, tab: var Table[U, V])
 
@@ -69,56 +68,7 @@ proc toNimVal[I, T](x: JlValue, locarr: var array[I, T]) =
   let x = toJlArray[T](x)
   toNimVal(x, locarr)
 
-  # Nim -> Julia
-# Avoid going throung template toJlVal pointer version when dealing with Julia known type
-# Is converter the right choice here ?
-converter nimValueToJlValue*[T](x: JlArray[T]): JlValue =
-  result = cast[JlValue](x)
-
-converter nimValueToJlValue(x: JlSym): JlValue =
-  result = cast[JlValue](x)
-
-converter nimValueToJlValue(x: JlDataType): JlValue =
-  result = cast[JlValue](x)
-
-converter nimValueToJlValue(x: JlFunc): JlValue  =
-  result = cast[JlValue](x)
-
-converter nimValueToJlValue(x: JlModule): JlValue  =
-  result = cast[JlValue](x)
-
-proc nimValueToJlValue*[T: SomeNumber|bool|pointer](val: T): JlValue =
-  result = jlBox(val)
-
-proc nimValueToJlValue(val: string): JlValue =
-  result = jlvalue_from_string(val)
-
-proc nimValueToJlValue(x: JlValue): JlValue  =
-  result = x
-
-# Complex stuff
-# Treat Nim object as Julia tuple
-# Forward declaration for cyclic import
-proc nimValueToJlValue(x: tuple): JlValue
-proc nimValueToJlValue(x: object): JlValue
-proc nimValueToJlValue[U, V](x: Table[U, V]): JlValue
-proc nimValueToJlValue[T](x: Option[T]): JlValue
-
-proc nimValueToJlValue[T](x: openarray[T]): JlValue  =
-  result = nimValueToJlValue(
-    jlArrayFromBuffer(x)
-  )
-
-proc nimValueToJlValue[I, T](x: array[I, T]): JlValue  =
-  result = nimValueToJlValue(
-    jlArrayFromBuffer(x)
-  )
-
-proc nimValueToJlValue[T](x: Tensor[T]): JlValue  =
-  ## Convert a Tensor to JlValue
-  result = nimValueToJlValue(
-    jlArrayFromBuffer(x)
-  )
+{.pop.}
 
 # Public API
 proc to*[U](x: JlArray[U], T: typedesc): T =
@@ -127,7 +77,6 @@ proc to*[U](x: JlArray[U], T: typedesc): T =
   else:
     toNimVal(x, result)
 
-
 proc to*(x: JlValue, T: typedesc): T =
   ## Copy a JlValue into a Nim type
   when T is void:
@@ -135,41 +84,15 @@ proc to*(x: JlValue, T: typedesc): T =
   else:
     toNimVal(x, result)
 
-proc toJlVal*[T](x: T): JlValue =
-  ## Convert a generic Nim type to a JlValue
-  nimValueToJlValue(x)
-
-proc toJlValue*[T](x: T): JlValue =
-  ## Alias for toJlVal
-  ## Added for consistency with JlValue type name
-  toJlVal[T](x)
-
 # Recursive import strategy
-import ./dicttuples
+import ./dict_tuples
+import ./obj_structs
+
+proc toNimVal(x: JlValue, t: var object) =
+  jlStructToNim(x, t)
 
 proc toNimVal(x: JlValue, t: var tuple) =
   jlTupleToNim(x, t)
 
 proc toNimVal[U, V](x: JlValue, tab: var Table[U, V]) =
   jlDictToNim[U, V](x, tab)
-
-proc nimValueToJlValue[T](x: Option[T]): JlValue  =
-  if isSome(x):
-    result = toJlVal(get(x))
-  else:
-    result = JlNothing
-
-proc nimValueToJlValue(x: tuple): JlValue  =
-  result = nimToJlTuple(x)
-
-# Treat Nim object as Julia tuple
-proc nimValueToJlValue(x: object): JlValue  =
-  result = nimToJlTuple(x)
-
-proc nimValueToJlValue[U, V](x: Table[U, V]): JlValue  =
-  result = nimTableToJlDict(x)
-
-{.pop.}
-
-export boxunbox
-export dicttuples