+++ tests/conversionstest.nim
@@ -19,16 +19,20 @@ proc tuplesTest() =
     check res.B == 2
     check res.C == 3
 
-  type TT = object
+proc objectTest() =
+  type MyStruct = object
     a: int
     b: Option[float]
     c: float
     d: Option[string]
-    e: Option[bool]
+    e: seq[int]
 
   block:
-    var tt: TT = TT(a: 123, b: some(-11.11e-3), c: 67.32147, d: some("azerty"), e: none(bool))
-    var ret = Julia.tupleTest(tt).to(bool)
+    var tt = MyStruct(a: 123, b: some(-11.11e-3), c: 67.32147, d: some("azerty"), e: @[1, 2, 3, 4, 5, 6])
+    var jltt = tt.toJlVal()
+    tt.e[0] = 111
+    check $(jltypeof(jltt)) == "MyStruct"
+    var ret = Julia.objectTest(jltt).to(bool)
     check ret
 
 proc stringModTest() =
@@ -82,9 +86,9 @@ proc dictToTableTest() =
     check res[key1] == val2
     check res[key2] == val1
 
-proc nestedObjectsTest() =
+proc nestedTuplesTest() =
   type
-    A = object
+    A = tuple
       dict : Table[string, float32]
       dat: seq[int]
 
@@ -93,22 +97,25 @@ proc nestedObjectsTest() =
       y: int
       z: int
 
-    O = object
+    O = tuple
       a: A
       b: B
 
   var o : O
-  o.a = A(dict: {"A": 1.0.float32,"B": 2.0.float32}.toTable, dat: toSeq(1..10))
+  o.a = (dict: {"A": 1.0.float32,"B": 2.0.float32}.toTable, dat: toSeq(1..10))
   o.b = (x: 36, y: 48, z: 60)
-  var res = Julia.nestedObjects(o)
+  var res = Julia.nestedTuples(o)
   check res.to(bool)
 
 proc runConversionsTest*() =
-  suite "Converions":
+  suite "Conversions":
     teardown: jlGcCollect()
 
     test "Tuples":
       tuplesTest()
+
+    test "Objects":
+      objectTest()
 
     test "String":
       stringModTest()
@@ -119,8 +126,8 @@ proc runConversionsTest*() =
     test "invertDict":
       dictToTableTest()
 
-    test "nestedObjectsTuples":
-      nestedObjectsTest()
+    test "nestedTuples":
+      nestedTuplesTest()
 
 when isMainModule:
   import ./testfull